Polishing - Reduce to single class KafkaJaasLoginModuleInitializer
